comparing changes to approved flow
